+++ dash/testing/application_runners.py
@@ -1,6 +1,7 @@
 from __future__ import print_function
 
 import sys
+import os
 import uuid
 import shlex
 import threading
@@ -68,6 +69,14 @@ class BaseDashRunner(object):
     def stop(self):
         raise NotImplementedError  # pragma: no cover
 
+    @staticmethod
+    def accessible(url):
+        try:
+            requests.get(url)
+        except requests.exceptions.RequestException:
+            return False
+        return True
+
     def __call__(self, *args, **kwargs):
         return self.start(*args, **kwargs)
 
@@ -90,6 +99,10 @@ class BaseDashRunner(object):
     def url(self):
         """the default server url"""
         return "http://localhost:{}".format(self.port)
+
+    @property
+    def is_windows(self):
+        return sys.platform == "win32"
 
 
 class ThreadedRunner(BaseDashRunner):
@@ -145,15 +158,8 @@ class ThreadedRunner(BaseDashRunner):
 
         self.started = self.thread.is_alive()
 
-        def accessible():
-            try:
-                requests.get(self.url)
-            except requests.exceptions.RequestException:
-                return False
-            return True
-
         # wait until server is able to answer http request
-        wait.until(accessible, timeout=1)
+        wait.until(lambda: self.accessible(self.url), timeout=1)
 
     def stop(self):
         requests.get("{}{}".format(self.url, self.stop_route))
@@ -180,7 +186,7 @@ class ProcessRunner(BaseDashRunner):
 
         args = shlex.split(
             "waitress-serve --listen=0.0.0.0:{} {}".format(port, entrypoint),
-            posix=sys.platform != "win32",
+            posix=not self.is_windows,
         )
         logger.debug("start dash process with %s", args)
 
@@ -188,6 +194,9 @@ class ProcessRunner(BaseDashRunner):
             self.proc = subprocess.Popen(
                 args, stdout=subprocess.PIPE, stderr=subprocess.PIPE
             )
+            # wait until server is able to answer http request
+            wait.until(lambda: self.accessible(self.url), timeout=3)
+
         except (OSError, ValueError):
             logger.exception("process server has encountered an error")
             self.started = False
@@ -214,3 +223,55 @@ class ProcessRunner(BaseDashRunner):
                 )
                 self.proc.kill()
                 self.proc.communicate()
+
+
+class RRunner(ProcessRunner):
+    def __init__(self, keep_open=False, stop_timeout=3):
+        super(RRunner, self).__init__(
+            keep_open=keep_open, stop_timeout=stop_timeout
+        )
+        self.proc = None
+
+    # pylint: disable=arguments-differ
+    def start(self, app):
+        """Start the server with waitress-serve in process flavor """
+
+        # app is a R string chunk
+        if not (os.path.isfile(app) and os.path.exists(app)):
+            path = (
+                "/tmp/app_{}.R".format(uuid.uuid4().hex)
+                if not self.is_windows
+                else os.path.join(
+                    (os.getenv("TEMP"), "app_{}.R".format(uuid.uuid4().hex))
+                )
+            )
+            logger.info("RRuner start => app is R code chunk")
+            logger.info("make a temporay R file for execution=> %s", path)
+            logger.debug("the content of dashR app")
+            logger.debug("%s", app)
+
+            with open(path, "w") as fp:
+                fp.write(app)
+
+            app = path
+
+        logger.info("Run dashR app with Rscript => %s", app)
+        args = shlex.split(
+            "Rscript {}".format(os.path.realpath(app)),
+            posix=not self.is_windows,
+        )
+        logger.debug("start dash process with %s", args)
+
+        try:
+            self.proc = subprocess.Popen(
+                args, stdout=subprocess.PIPE, stderr=subprocess.PIPE
+            )
+            # wait until server is able to answer http request
+            wait.until(lambda: self.accessible(self.url), timeout=2)
+
+        except (OSError, ValueError):
+            logger.exception("process server has encountered an error")
+            self.started = False
+            return
+
+        self.started = True