+++ tests/unit/test_app_runners.py
@@ -1,4 +1,3 @@
-import time
 import sys
 import requests
 import pytest
@@ -27,7 +26,6 @@ def test_threaded_server_smoke(dash_thread_server):
 )
 def test_process_server_smoke(dash_process_server):
     dash_process_server("simple_app")
-    time.sleep(2.5)
     r = requests.get(dash_process_server.url)
     assert r.status_code == 200, "the server is reachable"
     assert 'id="react-entry-point"' in r.text, "the entrypoint is present"